pyproject.toml
[tool.poetry]
name = "ecommerce-api"
version = "1.0.0"
description = "Production-ready ecommerce REST API"
authors = ["Your Team <team@example.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = { version = "^0.30.0", extras = ["standard"] }
gunicorn = "^22.0.0"
sqlalchemy = "^2.0.30"
alembic = "^1.13.0"
pyodbc = "^5.1.0"
python-dotenv = "^1.0.1"
pydantic = "^2.7.0"
pydantic-settings = "^2.3.0"
passlib = { version = "^1.7.4", extras = ["bcrypt"] }
bcrypt = "4.0.1"
python-jose = { version = "^3.3.0", extras = ["cryptography"] }
pyjwt = "^2.9.0"
slowapi = "^0.1.9"
structlog = "^24.1.0"
celery = "^5.4.0"
redis = "^5.0.4"
prometheus-client = "^0.20.0"
email-validator = "^2.1.1"
python-multipart = "^0.0.9"
faker = "^26.0.0"
numpy = "^2.0.0"
apscheduler = "^3.10.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^5.0.0"
httpx = "^0.27.0"
factory-boy = "^3.3.0"
ruff = "^0.4.0"
black = "^24.4.0"
mypy = "^1.10.0"
pre-commit = "^3.7.0"

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "I", "B", "UP"]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
strict = true
disallow_untyped_defs = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
